pyproject.toml
[tool.poetry]
name = "audgit"
version = "0.1.0"
description = ""
authors = ["Christopher David <chris@arcadelabs.co>", "Erik Aronesty <erik@q32.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
beautifulsoup4 = "^4.12.2"
pygithub = "^1.59.0"
python-dotenv = "^1.0.0"
anthropic = "^0.3.6"
nostr = {git = "https://github.com/earonesty/python-nostr.git", rev = "windows-compat"}

[tool.poetry.scripts]
audgit = "audgit.main:main"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
ruff = "^0.0.280"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
ignore = ['E701', 'E501']

[tool.ruff.per-file-ignores]
"__init__.py" = ['F401']

